+++ tests/default.nix
@@ -27,6 +27,7 @@ let
     // prefix "ep" (tryImport "endpoint.nix")
     // prefix "ln" (tryImport "listener.nix")
     // prefix "transport" (tryImport "transport.nix")
+    // prefix "hostname" (tryImport "hostname.nix")
     // prefix "registry" (tryImport "registry.nix")
     // prefix "iparse" (tryImport "internal/parse.nix")
     // prefix "ifmt" (tryImport "internal/format.nix")